Adds documentation support
Adds support for both manual and automatic API documentation to the
project. The chosen render engine was Sphinx. The API documentation is
generated with Doxygen on a XML that is later consumed by Breathe and
passed along to Sphinx for the final render.

Two Git submodules were added: breathe and sphinx_rtd_theme. The first
one is the bridge plugin for Sphinx that connects Sphinx and Doxygen.
The later is just a much better looking theme for the doc render. The
default theme looked pretty old.