Assistant for the generation of config.h files
CMake is very good tool, although it isn't that handy when it comes to
generate config.h files. By default the same check/variable/define has
to be defined in a number of different places, which besides tedious
is certainly prone to error.

gen-config.py assist with this task. It scans the CMakeLists.py file
an extract the headers, function and type size checks and generate the
config.h.in file that CMake will consume for the generation of the
final config.h file.