Simplifies how testcases are added
In the previous version of the 'check_add' macro if was necessary to
specify the name of the local variable that would store the testcase
information. Since it was something the macro could build itself there
was no point keeping this around. This new version uses a name based
on the name of the function that the tescase will execute.